refactor(structure): extract types to own files per project convention
