+++ MC1/Views/LineOfSight/LineOfSightLayoutMode.swift
@@ -1,0 +1,5 @@
+enum LineOfSightLayoutMode {
+    case map
+    case panel
+    case mapWithSheet
+}
