+++ MC1/Views/LineOfSight/LineOfSightViewModel.swift
@@ -15,6 +15,12 @@ enum PointID: Hashable {
     case repeater
 }
 
+// MARK: - PointID Identifiable Conformance
+
+extension PointID: Identifiable {
+    var id: Self { self }
+}
+
 // MARK: - Repeater Point
 
 /// A repeater point for relay analysis.
