+++ MC1/Views/Map/MapView.swift
@@ -286,255 +286,6 @@ struct MapView: View {
     }
 }
 
-// MARK: - Offline Badge
-
-private struct OfflineBadge: View {
-    var body: some View {
-        VStack {
-            HStack {
-                Text(L10n.Map.Map.OfflineBadge.label)
-                    .font(.caption)
-                    .bold()
-                    .padding(.horizontal)
-                    .padding(.vertical, 6)
-                    .background(.ultraThinMaterial, in: .capsule)
-                Spacer()
-            }
-            .padding(.leading)
-            Spacer()
-        }
-    }
-}
-
-// MARK: - Contact Detail Sheet
-
-private struct ContactDetailSheet: View {
-    let contact: ContactDTO
-    let onMessage: () -> Void
-    @Environment(\.dismiss) private var dismiss
-    @Environment(\.appState) private var appState
-
-    /// Sheet types for repeater flows
-    private enum ActiveSheet: Identifiable, Hashable {
-        case telemetryAuth
-        case telemetryStatus(RemoteNodeSessionDTO)
-        case adminAuth
-        case adminSettings(RemoteNodeSessionDTO)
-        case roomJoin
-
-        var id: String {
-            switch self {
-            case .telemetryAuth: "telemetryAuth"
-            case .telemetryStatus(let s): "telemetryStatus-\(s.id)"
-            case .adminAuth: "adminAuth"
-            case .adminSettings(let s): "adminSettings-\(s.id)"
-            case .roomJoin: "roomJoin"
-            }
-        }
-    }
-
-    @State private var activeSheet: ActiveSheet?
-    @State private var pendingSheet: ActiveSheet?
-
-    var body: some View {
-        NavigationStack {
-            List {
-                // Basic info section
-                Section(L10n.Map.Map.Detail.Section.contactInfo) {
-                    LabeledContent(L10n.Map.Map.Detail.name, value: contact.displayName)
-
-                    LabeledContent(L10n.Map.Map.Detail.type) {
-                        HStack {
-                            Image(systemName: typeIconName)
-                            Text(typeDisplayName)
-                        }
-                        .foregroundStyle(typeColor)
-                    }
-
-                    if contact.isFavorite {
-                        LabeledContent(L10n.Map.Map.Detail.status) {
-                            HStack {
-                                Image(systemName: "star.fill")
-                                Text(L10n.Map.Map.Detail.favorite)
-                            }
-                            .foregroundStyle(.orange)
-                        }
-                    }
-
-                    if contact.lastAdvertTimestamp > 0 {
-                        LabeledContent(L10n.Map.Map.Detail.lastAdvert) {
-                            ConversationTimestamp(date: Date(timeIntervalSince1970: TimeInterval(contact.lastAdvertTimestamp)), font: .body)
-                        }
-                    }
-                }
-
-                // Location section
-                Section(L10n.Map.Map.Detail.Section.location) {
-                    LabeledContent(L10n.Map.Map.Detail.latitude) {
-                        Text(contact.latitude, format: .number.precision(.fractionLength(6)))
-                    }
-
-                    LabeledContent(L10n.Map.Map.Detail.longitude) {
-                        Text(contact.longitude, format: .number.precision(.fractionLength(6)))
-                    }
-                }
-
-                // Path info section
-                Section(L10n.Map.Map.Detail.Section.networkPath) {
-                    if contact.isFloodRouted {
-                        LabeledContent(L10n.Map.Map.Detail.routing, value: L10n.Map.Map.Detail.routingFlood)
-                    } else {
-                        let hopCount = contact.pathHopCount
-                        LabeledContent(L10n.Map.Map.Detail.pathLength, value: hopCount == 1 ? L10n.Map.Map.Detail.hopSingular : L10n.Map.Map.Detail.hops(hopCount))
-                    }
-                }
-
-                // Actions section
-                Section {
-                    switch contact.type {
-                    case .repeater:
-                        Button {
-                            activeSheet = .telemetryAuth
-                        } label: {
-                            Label(L10n.Map.Map.Detail.Action.telemetry, systemImage: "chart.line.uptrend.xyaxis")
-                        }
-
-                        Button {
-                            activeSheet = .adminAuth
-                        } label: {
-                            Label(L10n.Map.Map.Detail.Action.management, systemImage: "gearshape.2")
-                        }
-
-                    case .room:
-                        Button {
-                            activeSheet = .roomJoin
-                        } label: {
-                            Label(L10n.Map.Map.Detail.Action.joinRoom, systemImage: "door.left.hand.open")
-                        }
-
-                    case .chat:
-                        Button {
-                            dismiss()
-                            onMessage()
-                        } label: {
-                            Label(L10n.Map.Map.Detail.Action.sendMessage, systemImage: "message.fill")
-                        }
-                        .radioDisabled(for: appState.connectionState)
-                    }
-                }
-            }
-            .navigationTitle(contact.displayName)
-            .navigationBarTitleDisplayMode(.inline)
-            .toolbar {
-                ToolbarItem(placement: .confirmationAction) {
-                    Button(L10n.Map.Map.Common.done) {
-                        dismiss()
-                    }
-                }
-            }
-            .sheet(item: $activeSheet, onDismiss: presentPendingSheet) { sheet in
-                switch sheet {
-                case .telemetryAuth:
-                    if let role = RemoteNodeRole(contactType: contact.type) {
-                        NodeAuthenticationSheet(
-                            contact: contact,
-                            role: role,
-                            customTitle: L10n.Map.Map.Detail.Action.telemetryAccessTitle
-                        ) { session in
-                            pendingSheet = .telemetryStatus(session)
-                            activeSheet = nil
-                        }
-                        .presentationSizing(.page)
-                    }
-
-                case .telemetryStatus(let session):
-                    RepeaterStatusView(session: session)
-
-                case .adminAuth:
-                    if let role = RemoteNodeRole(contactType: contact.type) {
-                        NodeAuthenticationSheet(contact: contact, role: role) { session in
-                            if session.isAdmin {
-                                pendingSheet = .adminSettings(session)
-                            } else {
-                                pendingSheet = .telemetryStatus(session)
-                            }
-                            activeSheet = nil
-                        }
-                        .presentationSizing(.page)
-                    }
-
-                case .adminSettings(let session):
-                    NavigationStack {
-                        RepeaterSettingsView(session: session)
-                            .toolbar {
-                                ToolbarItem(placement: .confirmationAction) {
-                                    Button(L10n.Map.Map.Common.done) {
-                                        activeSheet = nil
-                                    }
-                                }
-                            }
-                    }
-                    .presentationSizing(.page)
-
-                case .roomJoin:
-                    if let role = RemoteNodeRole(contactType: contact.type) {
-                        NodeAuthenticationSheet(contact: contact, role: role) { session in
-                            activeSheet = nil
-                            dismiss()
-                            appState.navigation.navigateToRoom(with: session)
-                        }
-                        .presentationSizing(.page)
-                    }
-                }
-            }
-        }
-    }
-
-    // MARK: - Sheet Management
-
-    private func presentPendingSheet() {
-        if let next = pendingSheet {
-            pendingSheet = nil
-            activeSheet = next
-        }
-    }
-
-    // MARK: - Computed Properties
-
-    private var typeIconName: String {
-        switch contact.type {
-        case .chat:
-            "person.fill"
-        case .repeater:
-            "antenna.radiowaves.left.and.right"
-        case .room:
-            "person.3.fill"
-        }
-    }
-
-    private var typeDisplayName: String {
-        switch contact.type {
-        case .chat:
-            L10n.Map.Map.NodeKind.chatContact
-        case .repeater:
-            L10n.Map.Map.NodeKind.repeater
-        case .room:
-            L10n.Map.Map.NodeKind.room
-        }
-    }
-
-    private var typeColor: Color {
-        switch contact.type {
-        case .chat:
-            .blue
-        case .repeater:
-            .green
-        case .room:
-            .purple
-        }
-    }
-}
-
 // MARK: - Preview
 
 #Preview("Map with Contacts") {
